InMemoryHistoryManager: класс Node реализовал как внутренний класс
