+++ src/controllers/HistoryManager.java
@@ -1,6 +1,7 @@
 package controllers;
-import model.Node;
+
 import model.Task;
+
 import java.util.ArrayList;
 
 public interface HistoryManager {
